Add : MainRight Component
