+++ public/data/commentData.json
@@ -1,5 +1,1 @@
-[
-  "Welcome to world best coding bootcamp!",
-  "Welcome to world best coding bootcamp!",
-  "Welcome to world best coding bootcamp!"
-]
+["Welcome to world best coding bootcamp!", "Hello!", "Amazing!"]
