+++ src/pages/hyunminChoi/Main/Main.scss
@@ -29,6 +29,7 @@
     /* feed 영역 */
     .feeds {
       width: 60%;
+      margin-bottom: 30px;
 
       /* 게시물 영역 */
       .content {
